requirements.txt
textblob
python-dotenv
flask_cors
flask
google.generativeai
yfinance
requests